Day 3: 쿠폰 목록 조회 API (#4)
* Day 3: 쿠폰 목록 조회 API

* [수정] Day 3: 쿠폰 목록 조회 API

- 테스트 오류 수정

* [수정] Day 3: 쿠폰 목록 조회 API 코드 개선

- 코드 간결화
- 불필요한 스트림 생성 개선
- 테스트 시 컴파일 시점에 필드 유효성 검사

* [수정] Day 3: 쿠폰 목록 조회 API 서비스 테스트 추가

* [수정] Day 3: 쿠폰 목록 조회 API 서비스 할인 금액 정밀도 개선

- BigDecimal로 변경하여 정밀도 개선

+++ src/main/java/example/dailycoding/coupon/dto/CouponDto.java
@@ -5,6 +5,7 @@ import example.dailycoding.coupon.domain.DiscountType;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Getter
@@ -13,7 +14,7 @@ public class CouponDto {
     private String id;
     private String name;
     private DiscountType discountType;
-    private double discount;
+    private BigDecimal discount;
     private LocalDateTime startDate;
     private LocalDateTime endDate;
 
@@ -22,7 +23,7 @@ public class CouponDto {
                 coupon.getId(),
                 coupon.getName(),
                 coupon.getDiscountType(),
-                coupon.getDiscount().longValue(),
+                coupon.getDiscount(),
                 coupon.getStartDate(),
                 coupon.getEndDate()
         );
